use utils in binaries
