+++ src/bin/utils/mod.rs
@@ -153,6 +153,32 @@ mod tests {
     use cairo_lang_sierra::ProgramParser;
 
     #[test]
+    fn test_check_compiler_path() {
+        // Define file, folder, and invalid paths for testing
+        let file_path = Path::new("src/bin/cairo-native-run.rs");
+        let folder_path = Path::new("src/bin");
+        let invalid_path = Path::new("src/non-existing-file.rs");
+
+        // Test when single_file is true and the path is a file
+        assert!(check_compiler_path(true, file_path).is_ok());
+
+        // Test when single_file is false and the path is a file
+        assert!(check_compiler_path(false, file_path).is_err());
+
+        // Test when single_file is true and the path is a folder
+        assert!(check_compiler_path(true, folder_path).is_err());
+
+        // Test when single_file is false and the path is a folder
+        assert!(check_compiler_path(false, folder_path).is_ok());
+
+        // Test when single_file is true and the path does not exist
+        assert!(check_compiler_path(true, invalid_path).is_err());
+
+        // Test when single_file is false and the path does not exist
+        assert!(check_compiler_path(false, invalid_path).is_err());
+    }
+
+    #[test]
     fn test_find_function() {
         // Parse a simple program containing a function named "Func2"
         let program = ProgramParser::new().parse("Func2@6() -> ();").unwrap();
